requirements.txt
fastapi==0.85.1
uvicorn==0.19.0
http3==0.6.7
requests==2.28.1
python-dotenv==0.21.0
nltk==3.7
